pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "temple-builder"
version = "0.0.1"
description = ""
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [
  { name = "Steven Mendoza" }
]
keywords = []

dependencies = [
  "bcrypt==4.3.0",
  "cffi==1.17.1",
  "click==8.2.1",
  "colorama==0.4.6",
  "cryptography==45.0.2",
  "dotenv==0.9.9",
  "paramiko==3.5.1",
  "pycparser==2.22",
  "PyNaCl==1.5.0",
  "python-dotenv==1.1.0",
  "PyYAML==6.0.2"
]

[project.urls]
Homepage = "https://github.com/StevenMZH/Temple-Builder.git"
Documentation = "https://github.com/StevenMZH/Temple-Builder.git"
Source = "https://github.com/StevenMZH/Temple-Builder.git"
Issues = "https://github.com/StevenMZH/Temple-Builder/issues"

[project.scripts]
temple = "temple.cli:main"

[tool.setuptools.packages.find]
exclude = ["tests*", "docs*", "workspace*", "venv*"]
